network.HTMLFetcher - add integration test with empty webpage
Use an embedded Jetty server to serve a hard coded resource file as
a response to all requests, then make a request and expect that
response.

This is a sloppy commit because we need to be checking that the
parse trees of the resultant HTML is identical, not that the strings
are identical. We need a cute hack to remove a valid but unusual
line break that WebKit inserts into the very sparse page.

Also the test class isn't factored well at all and isn't extensible.
However this is a decent start. If this test ever fails you've
probably not installed PhantomJS properly! However the debug trace
should tell you what's failing.

[1] from one of the Jetty maintainers was very useful; I've just
copied one of their techniques for how they test their code.

[1] http://stackoverflow.com/questions/19113359/sending-request-to-the-newly-created-jetty-server-for-testing-purposes